requirements.txt
fastapi==0.115.0
 uvicorn[standard]==0.29.0
 motor==3.4.0
 pydantic==2.9.2
 python-dotenv==1.0.1
 openai==1.30.0
 pytest==8.2.0
 httpx==0.27.0